+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - New plugin Photoswipe for interactive image galleries
  - Auto attribute `data-target-width` and `data-target-height` for links to images
  - Auto attribute `width` and `height` for images
  - Saving autocorrected HTML+ files (if is file AUTOCORRECT)